% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dseckw.R
\name{dseckw}
\alias{dseckw}
\title{The density function of the Secant Kumaraswamy Weibull probability distribution.}
\usage{
dseckw(x, a, b, c, lambda)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{a}{A parameter.}

\item{b}{B parameter.}

\item{c}{C parameter.}

\item{lambda}{Lambda parameter.}
}
\value{
A vector with n observations of the Secant Kumaraswamy Weibull distribution.
}
\description{
The density function of the Secant Kumaraswamy Weibull probability distribution.
}
\examples{
dseckw(1, 1, 1, 1, 1)
dseckw(1, 2, 2, 1, 1)
}

