% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hseckw.R
\name{hseckw}
\alias{hseckw}
\title{The hazard rate function of the Secant Kumaraswamy Weibull probability distribution.}
\usage{
hseckw(x, a, b, c, lambda)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{a}{A parameter.}

\item{b}{B parameter.}

\item{c}{C parameter.}

\item{lambda}{Lambda parameter.}
}
\value{
A vector with n observations of the Secant Kumaraswamy Weibull distribution.
}
\description{
The hazard rate function of the Secant Kumaraswamy Weibull probability distribution.
}
\examples{
hseckw(1, 1, 1, 1, 1)
hseckw(1, 3, 0.5, 2, 2)
}

