% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rseckw.R
\name{rseckw}
\alias{rseckw}
\title{Generates random deviates from a SecantKumaraswamyWeibull probability distribution.}
\usage{
rseckw(n, a, b, c, lambda)
}
\arguments{
\item{n}{Number of observations to be generated.}

\item{a}{A parameter.}

\item{b}{B parameter.}

\item{c}{C parameter.}

\item{lambda}{Lambda parameter.}
}
\value{
A vector with n observations of the SecantKumaraswamyWeibull distribution.
}
\description{
Generates random deviates from a SecantKumaraswamyWeibull probability distribution.
}
\examples{
rseckw(1000, 1, 1, 1, 1)
rseckw(1000, 2, 2, 1, 1)
}

