% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sseckw.R
\name{sseckw}
\alias{sseckw}
\title{The survival function of the SecantKumaraswamyWeibull probability distribution.}
\usage{
sseckw(x, a, b, c, lambda)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{a}{A parameter.}

\item{b}{B parameter.}

\item{c}{C parameter.}

\item{lambda}{Lambda parameter.}
}
\value{
A vector with n observations of the SecantKumaraswamyWeibull distribution.
}
\description{
The survival function of the SecantKumaraswamyWeibull probability distribution.
}
\examples{
sseckw(1, 1, 1, 1, 1)
sseckw(1, 3, 0.5, 2, 2)
}

